first pass at improving worker code + optimizing polling
