+++ src/views/channel-view/ChannelView.vue
@@ -72,6 +72,8 @@ export default defineComponent({
       handler: async function (to) {
         if (!to.params.channelId) return;
 
+        console.log("Adding new links worker");
+        this.linksWorker?.terminate();
         const { linksWorker } = await store.dispatch.loadExpressions({
           channelId: to.params.channelId,
         });
